Refactor phoneme presets

+++ HAARP/Util.cs
@@ -21,8 +21,8 @@ namespace HAARP
             return a * (1.0f - delta) + b * delta;
         }
 
-        public static float FactorToDecibels(float factor) => (float)Math.Log10(factor) * 10.0f;
+        public static double FactorToDecibels(double factor) => Math.Log10(factor) * 10.0;
 
-        public static float DecibelsToFactor(float dB) => (float)Math.Pow(10.0, dB / 10);
+        public static double DecibelsToFactor(double dB) => Math.Pow(10.0, dB / 10.0);
     }
 }
